[kernel] Add 'precision' delay timer based on the precision_timer routines
